refactor: Move database manager to repository package
- Moved database manager to the repository package.
- Updated import paths to reflect the new location.
- Renamed `ConnectSqlcWithPool` to `connectSqlcWithPool`.
- Added repository interfaces for users, movies, tvshows, watchlists.
- Added `RawSql` method to access raw SQL queries.
