+++ main.go
@@ -8,7 +8,7 @@ import (
 	"github.com/erkinov-wtf/movie-manager-bot/internal/config/app"
 	"github.com/erkinov-wtf/movie-manager-bot/internal/routes"
 	"github.com/erkinov-wtf/movie-manager-bot/internal/storage/cache"
-	"github.com/erkinov-wtf/movie-manager-bot/internal/storage/database"
+	"github.com/erkinov-wtf/movie-manager-bot/internal/storage/database/repository"
 	"github.com/erkinov-wtf/movie-manager-bot/internal/tmdb"
 	"github.com/erkinov-wtf/movie-manager-bot/pkg/encryption"
 	"github.com/erkinov-wtf/movie-manager-bot/pkg/workers"
@@ -26,7 +26,7 @@ func main() {
 	cfg := config.MustLoad()
 	tmdbClient := tmdb.NewClient(cfg)
 	log.Print("api client initialized")
-	db := database.MustLoadDb(cfg)
+	db := repository.MustLoadDb(cfg)
 	encryptor := encryption.NewKeyEncryptor(cfg.General.SecretKey)
 	cacheManager := cache.NewCacheManager(db, encryptor)
